+++ lib/image.ex
@@ -8961,7 +8961,7 @@ defmodule Image do
       iex> Image.map_pages image_2, &Image.crop(&1, 0, 150, 200, 200)
       {:error,
         "Image does not have a page-height header. Perhaps the image wasn't opened with the `pages: -1` option " <>
-        "or libvips wasn't built with libwebp-dev configured?"}
+        "or libvips wasn't built with libwebp-dev/libgif-dev configured?"}
 
   """
   @doc since: "0.38.0", subject: "Operation"
@@ -9018,7 +9018,7 @@ defmodule Image do
       {:error, _reason} ->
         {:error, "Image does not have a page-height header. " <>
           "Perhaps the image wasn't opened with the `pages: -1` option or " <>
-          "libvips wasn't built with libwebp-dev configured?"}
+          "libvips wasn't built with libwebp-dev/libgif-dev configured?"}
     end
   end