+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Image 0.38.0
 
-This is the changelog for Image version 0.38.0 released on ______, 2023.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-image/image/tags)
+This is the changelog for Image version 0.38.0 released on September 11th, 2023.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-image/image/tags)
 
 ### Bug Fixes
 
@@ -15,6 +15,8 @@ This is the changelog for Image version 0.38.0 released on ______, 2023.  For ol
 * Adds `Image.map_pages/2` to execute a function for each page in a `.webp` or `.gif` image then reassemble the pages into a new image. This can be used to crop all the frames of an animated image.  Thanks to @Mefgalm for the prompt to get this done. Eventually this function will be used by default for `Image.crop/5` and `Image.thumnail/3`.
 
 * Adds `Image.pages/1` to return the number of pages in an image. For normal images the result will be `1`. For animated images it will be the number of frames in the animation.
+
+* Adds support for cropping multi-page images (like animated `.webp` and `.gif` images) in `Image.crop/5`.
 
 * Adds `Image.Shape.circle/2`, `Image.Shape.ellipse/3` and `Image.Shape.line/5` as well as their `!` counterparts.